src/requirements.txt
pandas==1.3.4
google-cloud-storage==1.43.0
requests==2.26.0
flask==2.0.2
